feat: refactor shared context to example test path
