+++ test/acceptance/golium_test.go
@@ -29,6 +29,7 @@ import (
 	"github.com/Telefonica/golium/steps/redis"
 	s3steps "github.com/Telefonica/golium/steps/s3"
 	"github.com/Telefonica/golium/test/acceptance/steps/aggregate"
+	"github.com/Telefonica/golium/test/acceptance/steps/shared"
 	"github.com/cucumber/godog"
 )
 
@@ -57,6 +58,7 @@ func InitializeScenario(ctx context.Context, scenarioCtx *godog.ScenarioContext)
 		elasticsearch.Steps{},
 		s3steps.Steps{},
 		http.Steps{},
+		shared.Steps{},
 		aggregate.Steps{},
 	}
 	for _, stepsInitializer := range stepsInitializers {
